requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
Django==2.2.5
django-bootstrap4==1.1.1
django-braces==1.13.0
Pillow==6.2.1
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
